+++ Sources/VComponents/Components/Inputs/TextView/VTextViewState.swift
@@ -1,0 +1,15 @@
+//
+//  VTextViewState.swift
+//  VComponents
+//
+//  Created by Vakhtang Kontridze on 01.10.22.
+//
+
+import SwiftUI
+import VCore
+
+// MARK: - V Text View Internal State
+typealias VTextViewInternalState = GenericState_EnabledFocusedDisabled
+
+// MARK: - V Text View Button Internal State
+typealias VTextViewButtonInternalState = GenericState_EnabledPressedFocusedDisabled